Add email notifications at start and end of GED prep course
* Also move validation code into common location
